backend/requirements.txt
fastapi
uvicorn[standard]
youtube-transcript-api
openai
python-dotenv
